backend/pyproject.toml
[tool.poetry]
name = "number-game-backend"
version = "0.1.0"
description = "FastAPI 기반 숫자게임 웹 백엔드"
authors = ["sinbc <sinbc@example.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.30.0"
sqlmodel = "^0.0.19"
pydantic-settings = "^2.4.0"
python-jose = "^3.3.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
redis = "^5.0.1"
asyncpg = "^0.29.0"
aiosqlite = "^0.20.0"
alembic = "^1.13.1"
httpx = "^0.27.0"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
ruff = "^0.6.8"

[build-system]
requires = ["poetry>=1.8.0"]
build-backend = "poetry.masonry.api"

